Feat: 상세페이지 pick
